Rework of how characters are organized
Needed to facilitate attacking, item management, and ability management
for characters. Separates characters into a logical root, a state node,
and an 'instance' of the class held by the state.

The root 'character.tscn' and 'character.gd' are what other scripts
interact with.

The 'instance' scenes/scripts like 'basic_character.tscn' and
'basic_character.gd' deal with details like shaders, collider shapes,
etc.

Character state is held in the 'character_state.gd' and is meant to
be used as a struct more than a node in the scene graph.